Added a benchmark to compare list implementations
The benchmark also checks std::forward_list, although I had to get a
little creative to get it to behave as if it was sorted. The benchmark is
pretty bare bones but gives a decent idea of the general numbers. Here's
the results of one run with 100,000 nodes (all in milliseconds):

Times for std::list:
    Insert: 44209
    Find: 19903
    Delete: 44655
Times for LinkedList:
    Insert: 31509
    Find: 82957
    Delete: 93877
Times for SkipList:
    Insert: 352
    Find: 261
    Delete: 372

The speed of the skip list is obvious here. It looks like I need to
optimize the basic LinkedList though, I'm getting killed on find/delete.