Optimized LinkedList and SkipList
Removed size() (it's easy to add back in if needed), moved List to a
header file, reworked the Node hierarchy.

New results below (for 100,000 nodes, in ms):

Times for std::list:
    Insert: 45953
    Find: 19692
    Delete: 53144
Times for LinkedList:
    Insert: 36430
    Find: 116153
    Delete: 58546
Times for SkipList:
    Insert: 91
    Find: 70
    Delete: 81

Still not sure why my LinkedList find is so slow... maybe I'll move it
to a header and inline it.